// src/enums/class.enum.ts
export enum ClassStatus {
  Draft = 100,
  Published = 200,
  Deleted = 900,
}

export enum Difficulty {
  Beginner = 100,
  Intermediate = 200,
  Experienced = 300,
}

export enum ClassRepeat {
  Daily = 100,
  Weekly = 200,
  Monthly = 300,
  Custom = 400,
  NoRepeat = 500,
}

export enum FrequencyUnits {
  Day = 1,
  Week = 2,
  Month = 3,
}

export enum MonthlyReps {
  particular_date = 1,
  particular_day = 2,
  last_day = 3,
}